+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-17 10:34:01
+
+### Changed Files
+- `src/features/policies/PolicyForm.tsx`
+- `src/features/settings/products/components/ProductForm.tsx`
+- `src/services/commissions/CommissionCRUDService.ts`
+- `src/services/commissions/CommissionCalculationService.ts`
+- `src/services/commissions/termModifierUtils.ts`
+- `src/services/policies/policyService.ts`
+- `src/types/product.types.ts`
+
+### Statistics
+```
+ 7 files changed, 406 insertions(+), 8 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-17 09:45:18
 
 ### Changed Files